improve stubs and integration test
this commit contains two different improvements:

1.) a simple one that just adds a new class named CallableStub which can
represent a callback with calling magic on an object method (so far there
was only one for calling magic that was with a static class function).

2.) the other less self-explaining (but also simple) is about to allow to
include/require stubs file multiple times.

composer requires this as we're using class aliasing and PHP differs
between "\Class" and "Class" so for testing edge-cases the auto-loader
is triggered often and could result in a fatal error because of duplicate
function definition.

this has been fixed by only conditionally defining all stubs when not yet
exist.

additionally invalid callbacks haven't been tested earlier in integration.
these have been added to base testcase so that testcases that need to can
benefit from a diverse set of invalid callbacks.

these invalid ones were directly used to test for the required invalid
argument exception.